+++ src/iri/segment.rs
@@ -16,6 +16,7 @@ pub struct Segment<'a> {
 
 impl<'a> Segment<'a> {
 	/// The special dot segment `.` indicating the current directory.
+	#[inline]
 	pub fn current() -> Segment<'static> {
 		Segment {
 			data: &[0x2e],
@@ -24,6 +25,7 @@ impl<'a> Segment<'a> {
 	}
 
 	/// The special dot segment `..` indicating the parent directory.
+	#[inline]
 	pub fn parent() -> Segment<'static> {
 		Segment {
 			data: &[0x2e, 0x2e],
@@ -32,15 +34,18 @@ impl<'a> Segment<'a> {
 	}
 
 	/// Get the length of the path name.
+	#[inline]
 	pub fn len(&self) -> usize {
 		self.data.len()
 	}
 
+	#[inline]
     pub fn as_ref(&self) -> &[u8] {
 		self.data
 	}
 
     /// Get the underlying segment slice as a string slice.
+	#[inline]
 	pub fn as_str(&self) -> &str {
 		unsafe {
 			std::str::from_utf8_unchecked(self.data)
@@ -48,6 +53,7 @@ impl<'a> Segment<'a> {
 	}
 
 	/// Get the underlying segment slice as a string slice by consuming the segment reference.
+	#[inline]
 	pub fn into_str(self) -> &'a str {
 		unsafe {
 			std::str::from_utf8_unchecked(self.data)
@@ -55,22 +61,26 @@ impl<'a> Segment<'a> {
 	}
 
     /// Get the underlying segment slice as a percent-encoded string slice.
+	#[inline]
 	pub fn as_pct_str(&self) -> &PctStr {
 		unsafe {
 			PctStr::new_unchecked(self.as_str())
 		}
 	}
 
+	#[inline]
 	pub fn is_open(&self) -> bool {
 		self.open
 	}
 
     /// Checks if the segment is empty.
+	#[inline]
 	pub fn is_empty(&self) -> bool {
 		self.data.is_empty()
 	}
 
 	/// Open that path.
+	#[inline]
 	pub fn open(&mut self) {
 		self.open = true
 	}
@@ -79,6 +89,7 @@ impl<'a> Segment<'a> {
 impl<'a> TryFrom<&'a str> for Segment<'a> {
 	type Error = Error;
 
+	#[inline]
 	fn try_from(str: &'a str) -> Result<Segment<'a>, Error> {
 		let segment_len = parsing::parse_path_segment(str.as_ref(), 0)?;
 		let data: &[u8] = str.as_ref();
@@ -101,6 +112,7 @@ impl<'a> TryFrom<&'a str> for Segment<'a> {
 }
 
 impl<'a> fmt::Display for Segment<'a> {
+	#[inline]
 	fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
 		if self.open {
 			write!(f, "{}/", self.as_str())
@@ -111,6 +123,7 @@ impl<'a> fmt::Display for Segment<'a> {
 }
 
 impl<'a> fmt::Debug for Segment<'a> {
+	#[inline]
 	fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
 		if self.open {
 			write!(f, "{}/", self.as_str())
@@ -121,6 +134,7 @@ impl<'a> fmt::Debug for Segment<'a> {
 }
 
 impl<'a> cmp::PartialEq for Segment<'a> {
+	#[inline]
 	fn eq(&self, other: &Segment) -> bool {
 		self.open == other.open && self.as_pct_str() == other.as_pct_str()
 	}
@@ -129,24 +143,28 @@ impl<'a> cmp::PartialEq for Segment<'a> {
 impl<'a> Eq for Segment<'a> { }
 
 impl<'a> cmp::PartialEq<&'a str> for Segment<'a> {
+	#[inline]
 	fn eq(&self, other: &&'a str) -> bool {
 		self.as_pct_str() == *other
 	}
 }
 
 impl<'a> PartialOrd for Segment<'a> {
+	#[inline]
 	fn partial_cmp(&self, other: &Segment<'a>) -> Option<Ordering> {
 		Some(self.cmp(other))
 	}
 }
 
 impl<'a> Ord for Segment<'a> {
+	#[inline]
 	fn cmp(&self, other: &Segment<'a>) -> Ordering {
 		self.as_pct_str().cmp(other.as_pct_str())
 	}
 }
 
 impl<'a> Hash for Segment<'a> {
+	#[inline]
 	fn hash<H: Hasher>(&self, hasher: &mut H) {
 		self.as_pct_str().hash(hasher)
 	}